Shows days in duration strings; try out JS testing
The original purpose of this commit was simply to show "10d 0h 5m" instead of "240h 5m".

But the first attempt at testing needed to use Playwright, so I added tooling to simplify JS
testing, which will also have the nice side effect of separating HTML and JS code blocks.